refactor: Split the settings view and group its alert arguments
SettingsView.swift was over the size the linter allows, and the
alert helper took eight arguments. The bindings and the actions now
travel as one value each.
